dev-tbot-76 > python: tests - moved fixtures to conftest
